Now, implementation is functional.
Now with this form, you can receive emails from visitors on your website who submit them. Cool right. The only set back now is the fact that your company is not 100% displayed to you and to the public by these services since FORMSUBMIT will like to advertise too. You will like to maximize your chances of getting clients right. If you like to help me with that, feel free to demand to be a contributor.